attempting to work around an anchoring problem
Nameplate aura anchors sometimes seem to start from the top center and grown to the right, even though they should have been vertically centered. 

This is not a consistent bug, so my theory is that centered anchoring points kind of "stick" to there position from when they were set, if the frame at that point was lacking a proper size. Maybe. Ish.