moving hotkey abbreviations to the actionbar module
We didn't actually want these localized, so having them in the locale file was my bad. 
We are however working on a system to display special keys like the arrows, page up/down, enter, tab and so on as icons instead of text, thus making them far easier to understand, fit them more nicely on the buttons, and avoid the need for localization! :)